refactor(webhook): Move function and update API path
