+++ raven/__init__.py
@@ -1,1 +1,3 @@
 __version__ = "2.2.2"
+
+from raven.raven_integrations.doctype.raven_incoming_webhook.raven_incoming_webhook import handle_incoming_webhook as webhook  # noqa
